Move tests and suites to different packages
Reusing validation and e2e tests in other suites was difficult due to
the suite and tests being in one package and flags being defined in the suites
init funtion ("junit", "report"). If trying to reuse the tests in a suite also
using these flags, an error will result.

(cherry picked from commit ec6dd26b3a1ecb607fe7a2ffb6a921d684c98935)
