+++ Makefile
@@ -55,12 +55,14 @@ all: manager ## Default make target if no options specified
 test-validation: generate fmt vet manifests  ## Run validation tests
 	rm -rf ${VALIDATION_TESTS_REPORTS_PATH}
 	mkdir -p ${VALIDATION_TESTS_REPORTS_PATH}
-	USE_LOCAL_RESOURCES=true go test --tags=validationtests -v ./test/validation -ginkgo.v -junit $(VALIDATION_TESTS_REPORTS_PATH) -report $(VALIDATION_TESTS_REPORTS_PATH)
+	USE_LOCAL_RESOURCES=true go test --tags=validationtests -v ./test/e2e/validation -ginkgo.v -junit $(VALIDATION_TESTS_REPORTS_PATH) -report $(VALIDATION_TESTS_REPORTS_PATH)
 
-test-e2e: generate fmt vet manifests  ## Run e2e tests
+test-functional: generate fmt vet manifests  ## Run e2e tests
 	rm -rf ${TESTS_REPORTS_PATH}
 	mkdir -p ${TESTS_REPORTS_PATH}
-	USE_LOCAL_RESOURCES=true go test --tags=e2etests -v ./test/e2e -ginkgo.v -junit $(TESTS_REPORTS_PATH) -report $(TESTS_REPORTS_PATH)
+	USE_LOCAL_RESOURCES=true go test --tags=e2etests -v ./test/e2e/functional -ginkgo.v -junit $(TESTS_REPORTS_PATH) -report $(TESTS_REPORTS_PATH)
+
+test-e2e: generate fmt vet manifests test-validation test-functional  ## Run e2e tests
 
 manager: generate fmt vet  ## Build manager binary
 	go build -ldflags "-X main.build=$$(git rev-parse HEAD)" -o bin/manager main.go
