--- test/e2e/e2e_test.go
@@ -1,11 +1,0 @@
-// +build e2etests
-
-package e2e
-
-import (
-	"testing"
-)
-
-func TestE2E(t *testing.T) {
-	RunE2ETests(t)
-}
