--- test/validation/validation_test.go
@@ -1,11 +1,0 @@
-// +build validationtests
-
-package validation
-
-import (
-	"testing"
-)
-
-func TestValidation(t *testing.T) {
-	RunValidationTests(t)
-}
